[ROCm] Fix for ROCm CSB Breakage - 200902 - 1
The following commit introduces regressions in two unit-tests, in the ROCm TF build

https://github.com/tensorflow/tensorflow/commit/458d0906ebe75072ebd87cf5088772405d0cd03e

```
//tensorflow/python/kernel_tests:conv_ops_3d_test_gpu                    FAILED in 3 out of 3 in 212.4s
//tensorflow/python/kernel_tests:conv_ops_test_gpu                       FAILED in 3 out of 3 in 591.0s
```

The failures are within subtests that test padding for the float16 type.

The cause of failure is a strange one. In the ROCm TF build,
* the call(s) to the `PadInput` functor are in the files (conv_*.cc) that are compiled by the "CPU" compiler, while the
++ the GPUDevice specific template instantiations of the  `PadInput` functor are in the files that are compiled by the "GPU" compiler.

For T == Eigen::half, the value of the "padding_value" argument (added by the commit above) seems to be getting corrupted, leading to regressions in the convolution unit tests.

The value seems to be correct when printed (via VLOG) at the call site(s), and corrupted when printed in the GPU specialization within `conv_2d_gpu.h`

I do not understand the exact reason for the this, but based on similar past issues, it is likely due to ABI incompatibility between the "old" CPU compiler (gcc 5.4 for Ubuntu 16.04, gcc 7.5 for Ubuntu 18.04) and the "new" ROCm GPU compiler (hipclang which is based on latest clang)

Adding an extra int argument to the functor seems to alleviate this issue, and results in the value for the "padding_value" argument no longer getting corrupted.

This workaround can be removed once the "CPU" compiler is gcc8 or higher.